requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
infobip-api-python-sdk==5.0.1
install==1.3.5
multidict==6.0.4
Pillow==9.5.0
psycopg2==2.9.6
pydantic==1.10.9
PyJWT==2.6.0
pytz==2023.3
regex==2022.3.15
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.2.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
gunicorn==20.1.0
